requirements.txt
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
Django==2.1.1
django-storages==1.9.1
docutils==0.15.2
jmespath==0.9.5
numpy==1.18.2
Pillow==7.1.1
python-dateutil==2.8.1
pytz==2020.1
rawkit==0.6.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
